chore(build): add post deployment script for azure assemblies
Because this deployment keeps the `App_Data\Sites\[tenant]\settings.txt` files
in Azure blob storage instead of the file system, it requires certain assemblies
and config files from the **Orchard.Azure** module to be located in the root of
the codebase.

This fix removes those files & adds a post deployment script to be run that does this copy on deployments automatically. To run this, add a `POST_DEPLOYMENT_ACTION` app setting on the web app pointing to the location of the deployment script.

Closes #8.